refactor correctly for Java language
