--- java-code-challenges/src/main/java/java/code/challenges/ArrayShift.java
@@ -1,7 +1,0 @@
-package java.code.challenges;
-
-public class ArrayShift {
-    public static void main(String[] args) {
-
-    }
-}
