+++ java-code-challenges/src/main/java/javachallenges/Library.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package java.code.challenges;
+package javachallenges;
 
 public class Library {
     public boolean someLibraryMethod() {
