--- java-code-challenges/src/test/java/java/code/challenges/ArrayShiftTest.java
@@ -1,15 +1,0 @@
-package java.code.challenges;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-
-public class ArrayShiftTest {
-    public class LibraryTest {
-        @Test
-        public void testSomeLibraryMethod() {
-            Library classUnderTest = new Library();
-            assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
-        }
-    }
-}
