+++ java-code-challenges/src/test/java/javachallenges/ArrayShiftTest.java
@@ -1,0 +1,20 @@
+package javachallenges;
+
+import org.junit.Test;
+
+import java.lang.reflect.Array;
+
+import static org.junit.Assert.assertTrue;
+
+public class ArrayShiftTest {
+    @Test
+    public void testSomeLibraryMethod() {
+        Library classUnderTest = new Library();
+        assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
+    }
+//    @Test
+//    public void testArrayTest {
+//        ArrayShift whatWeAreTesting = new ArrayShift();
+//
+//    }
+}
