+++ java-code-challenges/src/test/java/javachallenges/LibraryTest.java
@@ -1,13 +1,14 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package java.code.challenges;
+package javachallenges;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
 
 public class LibraryTest {
-    @Test public void testSomeLibraryMethod() {
+    @Test
+    public void testSomeLibraryMethod() {
         Library classUnderTest = new Library();
         assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
     }
